new movements of the queen and the bishop by inheriting a class
